FlyingDart corregido movimiento

+++ gdx-logic-bricks/src/main/java/com/indignado/logicbricks/core/actuators/InstanceEntityActuator.java
@@ -11,7 +11,7 @@ public class InstanceEntityActuator<T extends Entity> extends Actuator {
     public EntityFactory entityFactory;
     public Vector2 localPosition;
     public float angle = 0;
-    public short duration = 0; // 0 = lives forever
+    public float duration = 0; // 0 = lives forever
     public Type type;
 
     public enum Type {AddEntity, RemoveEntity}
