+++ gdx-logic-bricks/src/main/java/com/indignado/logicbricks/core/sensors/TimerSensor.java
@@ -9,7 +9,7 @@ public class TimerSensor extends Sensor {
     public static short ALWAYS = 0;
     public boolean repeat = false;
     public short delay = 0;
-    public short duration = ALWAYS;
+    public float duration = ALWAYS;
 
 
 }
